refacotr: 공통되는 코드 분리
